+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Playlist Tools\Playlist Manipulation\Harmonic mix: cleanup of some menu entries and code logic.
 - Playlist Tools\Selection Manipulation\Advanced sort: cleanup of some menu entries and code logic.
 - Playlist Tools\Presets: are now saved at '.profile/js_data/export' folder when exported.
+- Playlist Tools\Presets: individual entries can now be exported at the 'Edit entries from list' submenus as long as they are not the default ones.
 - UI: updated icons for Playlist Tools submenu custom button for all new and renamed menus.
 - Fonts: fonts are no longer bundled at '_resources' folder, but found at: https://github.com/regorxxx/foobar2000-assets/tree/main/Fonts
 - Binaries: external binaries are no longer bundled along scripts, but found at: https://github.com/regorxxx/foobar2000-assets/tree/main/binaries